pipe_macro: Make IFOComboMap index with bits indicating which IFOs are being used
Previously IFOComboMap was indexed as [single detector, two detectors,
three detectors, ...]. This means that checking to see how many and
which detectors are in use can't be easily done without referring back
to the table, and any changes to the table can potentially break other
parts of the code.

This patch changes IFOComboMap to be indexed by powers of two. That
means that any index of (2^i - 1) is a single detector, indicated by
index into IFOMap. Now when using IFOComboMap, the number of detectors
in use can be determined using `__builtin_popcount(icombo + 1)`, and
whether a single detector is in use can be done with the check `index &
(icombo + 1)`.