cohfar: Reduce the size of outputted XML files by only using active IFOs
Previously, every possible combination of existing IFO was outputted as
a part of the XML dump from `cohfar`. This meant that the size of the
XML dump wasn't tied to the actual number of IFOs in use, and would
continue to exponentially increase in size as we increased the number of
IFOs that might be in play (e.g. KAGRA, LIGO-India, etc.).

To solve this, we have moved to using the total combination (as an index
into IFOComboMap) as a store of the active IFOs, and using
`__builtin_popcount()` to get the total number of IFOs that are in use
in that combination. This means that we can quickly and efficiently
determine the maximum number of detectors that we actually need to print
stats for, with no additional information required.

-- Tom Almeida