postcohtable: Remove detector names from table
This change removes detector names from the various data structures that
are used throughout the pipeline and ensures that all pointers into the
data structures have their offsets based upon the number of
interferometers.

Previously, adding support for an additional interferometer to the
pipeline would involve modifying all of the below files, and would often
require extensive testing to ensure that the massive number of changes
that would need to be done would be correct.

With these changes, when adding interferometers, the only file that
needs to be changed is `gstlal-spiir/include/pipe_macro.h`, with the
modification of `MAX_NIFO`, `IFOMap` and `IFOComboMap`. By changing
these variables, the next compilation of the pipeline will automatically
support the new interferometer.

-- Tom Almeida