Rename add_psd_initial_coinc to append_psd_to_coincs_doc
This boolean has been renamed to be slightly more clear about
what functionality this condition is actually handling.
Append is used as we are typically inserting this LIGOLW element
to the end of the list of LIGOW elements in the LIGOLW XML document,
rather than performing any kind of mathematical addition.